mods_panel: create mods_panel_icons.py and move class ModListIcons to it
