+++ app/views/mods_panel.py
@@ -60,6 +60,7 @@ from app.views.dialogue import (
     show_dialogue_input,
     show_warning,
 )
+from app.views.mods_panel_icons import ModListIcons
 
 
 def uuid_no_key(uuid: str) -> str:
@@ -473,83 +474,6 @@ class ModListItemInner(QWidget):
             self.main_label.setObjectName(new_widget_object_name)
             self.main_label.style().unpolish(self.main_label)
             self.main_label.style().polish(self.main_label)
-
-
-class ModListIcons:
-    _data_path: Path = AppInfo().theme_data_folder / "default-icons"
-    _ludeon_icon_path: str = str(_data_path / "ludeon_icon.png")
-    _local_icon_path: str = str(_data_path / "local_icon.png")
-    _steam_icon_path: str = str(_data_path / "steam_icon.png")
-    _csharp_icon_path: str = str(_data_path / "csharp.png")
-    _xml_icon_path: str = str(_data_path / "xml.png")
-    _git_icon_path: str = str(_data_path / "git.png")
-    _steamcmd_icon_path: str = str(_data_path / "steamcmd_icon.png")
-    _warning_icon_path: str = str(_data_path / "warning.png")
-    _error_icon_path: str = str(_data_path / "error.png")
-
-    _ludeon_icon: QIcon | None = None
-    _local_icon: QIcon | None = None
-    _steam_icon: QIcon | None = None
-    _csharp_icon: QIcon | None = None
-    _xml_icon: QIcon | None = None
-    _git_icon: QIcon | None = None
-    _steamcmd_icon: QIcon | None = None
-    _warning_icon: QIcon | None = None
-    _error_icon: QIcon | None = None
-
-    @classmethod
-    def ludeon_icon(cls) -> QIcon:
-        if cls._ludeon_icon is None:
-            cls._ludeon_icon = QIcon(cls._ludeon_icon_path)
-        return cls._ludeon_icon
-
-    @classmethod
-    def local_icon(cls) -> QIcon:
-        if cls._local_icon is None:
-            cls._local_icon = QIcon(cls._local_icon_path)
-        return cls._local_icon
-
-    @classmethod
-    def steam_icon(cls) -> QIcon:
-        if cls._steam_icon is None:
-            cls._steam_icon = QIcon(cls._steam_icon_path)
-        return cls._steam_icon
-
-    @classmethod
-    def csharp_icon(cls) -> QIcon:
-        if cls._csharp_icon is None:
-            cls._csharp_icon = QIcon(cls._csharp_icon_path)
-        return cls._csharp_icon
-
-    @classmethod
-    def xml_icon(cls) -> QIcon:
-        if cls._xml_icon is None:
-            cls._xml_icon = QIcon(cls._xml_icon_path)
-        return cls._xml_icon
-
-    @classmethod
-    def git_icon(cls) -> QIcon:
-        if cls._git_icon is None:
-            cls._git_icon = QIcon(cls._git_icon_path)
-        return cls._git_icon
-
-    @classmethod
-    def steamcmd_icon(cls) -> QIcon:
-        if cls._steamcmd_icon is None:
-            cls._steamcmd_icon = QIcon(cls._steamcmd_icon_path)
-        return cls._steamcmd_icon
-
-    @classmethod
-    def warning_icon(cls) -> QIcon:
-        if cls._warning_icon is None:
-            cls._warning_icon = QIcon(cls._warning_icon_path)
-        return cls._warning_icon
-
-    @classmethod
-    def error_icon(cls) -> QIcon:
-        if cls._error_icon is None:
-            cls._error_icon = QIcon(cls._error_icon_path)
-        return cls._error_icon
 
 
 class ModListWidget(QListWidget):
